Always copy bootable images in BOOTABLE_IMAGES/
Images placed in BOOTABLE_IMAGES/ are not re-generated when signing
the target-files zip content. As of now, images are put into
BOOTABLE_IMAGES/ only if BOARD_CUSTOM_BOOTIMG is set to true. Remove
the flag and copy the images unconditionally since there's no need
to re-generate them now that ro.bootimage.build.fingerprint is inside
the build.prop.

Change-Id: Idbb5329e0598c04c39fe7f1899b6af3dbdd4f0aa
Signed-off-by: Louis Popi <theh2o64@gmail.com>